+++ Webresource.Syncer/Webresource.Syncer/Helper/Map.cs
@@ -1,6 +1,6 @@
-﻿using Webresource.Syncer.Enum;
+﻿using WebResource.Syncer.Enum;
 
-namespace Webresource.Syncer.Helper
+namespace WebResource.Syncer.Helper
 {
     class Map
     {